Modify anchors list in GeospatialActivity to be thread safe
Modify anchors list so that it is a CopyOnWriteArrayList instead of a
Kotlin mutableList. The number of writes to this list is expected to be
small (any time an anchor is added or all anchors are cleared) versus
the number of reads (every frame), so the performance hit for list
modifications is expected to be minimal.

Bug: 414825430
Test: Using Whitebox GeospatialActivity on Pixel 9 Pro

Change-Id: I5ee4f5f3f67b6cd2e7350fcd8e6a0d97e259d75a