Add web targets to room-testing
To support web the main two testing APIs, createDatabase() and runMigrationsAndValidate() are now suspend functions. The actual declaration of the MigrationTestHelper for web is defined but not yet implemented.

This change is to allow dependants to target web. Reading schema files for tests that target web is still pending.

Bug: 336758416
Bug: 438041176
Test: :room3:room3-testing:allTests
Relnote: TODO
Change-Id: I2d285e1c0fc8b79fe2b944088b6643cac889931c