Add trackpad gesture handling
Relnote: "Adds support for trackpad gestures and conversion from trackpad
  input events to mouse.
  With this change, trackpad pointer events that control a cursor like on
  tablets or laptops will generally be reported as mouse pointers.
  Pan and scale gestures will also be reported with additional information
  available in the pointer event changes, with a new PointerEventType.Pan
  and PointerEventType.Scale indicating that these values will be set"

These changes are guarded behind the feature flag of
`ComposeUiFlags.isTrackpadGestureHandlingEnabled` and
`ComposeFoundationFlags.isTrackpadGestureHandlingEnabled`.
For ease of reverting, in this CL the flags are set to `false`,
but in the follow-up CL, the flags are flipped to `true` by default.

Test: New tests
Bug: 315527861
Fixes: 459831570

Change-Id: Id071af573d5710d7d3aca73bd1c8037bdb0ea701