`XrStubs.get` now returns a default instance instead of `null` when not installed
This default instance is less error-prone to version mismatches between the provided stub library on the classpath vs the stub library that M3 is compiled against.

Without a default instance, if a new stub is added and we forget to add a default return value for it, then a crash could occur when an older version of the stubs library is on the classpath, as that function was never implemented.

With a default instance, every new function must be implemented by the `NoOpXrStubs` implementation, thus these crashes will not be possible.

Change-Id: I905e24072381e3978f69777e7c01b28b6a6a6964
Test: Builds and existing tests pass
Relnote: `XrStubs.get` now returns a default instance instead of `null` when not installed